chore(exchange): do not register entities on a separate connection
